paper-explainer/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pydantic>=2.0.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
